brief 159: int64-widen CP9_MX/CP9_FMX to fix genome-scale int32 overflow
The non-banded CP9 HMM matrix is (N+1)*(M+1) cells. At genome scale
(N=L~197k, M=clen~197k for MPXV) that is ~3.9e10 cells, 18x over the
int32 max. GrowCP9Matrix()/CreateCP9Matrix() computed the cell count and
the per-row pointer offsets (i*(M+1)) in int32, so the count wrapped to
~2.45e8: a tiny (~980 MB x4) allocation succeeded (peak RSS ~6 GB, NOT an
OOM) and the row pointers were set far out of bounds -> SIGSEGV (signal
11) when cp9_Forward/Backward filled the matrix.

This path is reached by cmalign -g --p7band on a genome target: the
p7-banded checkpointed CP9 band derivation returns eslERANGE (467 GB
alignment matrix > --mxsize), and DispatchSqAlignment falls back to a
standard *non-banded* full CP9 F/B (cp9_Forward, nrows=L), which hits the
overflow. This is the band-derivation CP9 matrix; the truncated alignment
matrix (cm_mx.c cm_tr_hb_mx_SizeNeeded/GrowTo) was already int64-clean.

Fix (int64 analogue of brief 147, minimal/targeted):
- CP9_MX, CP9_FMX: ncells_allocated, ncells_valid int -> int64_t
- Create/Grow/SizeNeeded CP9Matrix + CP9FMatrix mirrors: compute
  (N+1)*(M+1) and i*(M+1) and the running cell offset in int64
- Initialize{,F}CP9Matrix: cast ncells_valid to int for esl_vec_{I,F}Set
  (allocated CP9 mxes are always < 2^31 cells; genome-scale non-banded
  mxes fail ESL_ALLOC before Initialize is reached)

Sub-2^31 results are numerically identical (wider type, same values), so
all sub-genome trunc/non-trunc/int paths are byte-identical to 7e8a3647.
At genome scale the 155 GB non-banded F/B now fails as a clean eslEMEM
(no signal 11) under h_vmem; completion still needs the --ckpt
composition (brief 158).

+++ src/infernal.h
@@ -733,8 +733,8 @@ typedef struct cp9_mx_s {
   /* variables added for HMMER3 p7 HMM banding of CP9 HMM dp algorithms */
   int *kmin;            /* OPTIONAL (can be null) [0.1..i..rows] = k, minimum node for residue i is k */
   int *kmax;            /* OPTIONAL (can be null) [0.1..i..rows] = k, maximum node for residue i is k */
-  int  ncells_allocated; /* number of cells allocated in matrix */
-  int  ncells_valid;     /* number of cells currently valid in the matrix */
+  int64_t ncells_allocated; /* number of cells allocated in matrix (int64: a non-banded genome-scale CP9 mx is (L+1)*(M+1) ~ 4e10 cells, overflows int32) */
+  int64_t ncells_valid;     /* number of cells currently valid in the matrix (int64, see ncells_allocated) */
 
 } CP9_MX;
 
@@ -764,8 +764,8 @@ typedef struct cp9_fmx_s {
   /* variables added for HMMER3 p7 HMM banding of CP9 HMM dp algorithms */
   int *kmin;            /* OPTIONAL (can be null) [0.1..i..rows] = k, minimum node for residue i is k */
   int *kmax;            /* OPTIONAL (can be null) [0.1..i..rows] = k, maximum node for residue i is k */
-  int  ncells_allocated; /* number of cells allocated in matrix */
-  int  ncells_valid;     /* number of cells currently valid in the matrix */
+  int64_t ncells_allocated; /* number of cells allocated in matrix (int64, see CP9_MX) */
+  int64_t ncells_valid;     /* number of cells currently valid in the matrix (int64, see CP9_MX) */
 
 } CP9_FMX;
